migrate swap feature to xredis_2.

+++ tests/unit/multi.tcl
@@ -212,43 +212,45 @@ start_server {tags {"multi"}} {
         r exec
     } {PONG}
 
-    test {SWAPDB is able to touch the watched keys that exist} {
-        r flushall
-        r select 0
-        r set x 30
-        r watch x ;# make sure x (set to 30) doesn't change (SWAPDB will "delete" it)
-        r swapdb 0 1
-        r multi
-        r ping
-        r exec
-    } {}
+    if {!$::swap} {
+        test {SWAPDB is able to touch the watched keys that exist} {
+            r flushall
+            r select 0
+            r set x 30
+            r watch x ;# make sure x (set to 30) doesn't change (SWAPDB will "delete" it)
+            r swapdb 0 1
+            r multi
+            r ping
+            r exec
+        } {}
 
-    test {SWAPDB is able to touch the watched keys that do not exist} {
-        r flushall
-        r select 1
-        r set x 30
-        r select 0
-        r watch x ;# make sure the key x (currently missing) doesn't change (SWAPDB will create it)
-        r swapdb 0 1
-        r multi
-        r ping
-        r exec
-    } {}
+        test {SWAPDB is able to touch the watched keys that do not exist} {
+            r flushall
+            r select 1
+            r set x 30
+            r select 0
+            r watch x ;# make sure the key x (currently missing) doesn't change (SWAPDB will create it)
+            r swapdb 0 1
+            r multi
+            r ping
+            r exec
+        } {}
 
-    test {WATCH is able to remember the DB a key belongs to} {
-        r select 5
-        r set x 30
-        r watch x
-        r select 1
-        r set x 10
-        r select 5
-        r multi
-        r ping
-        set res [r exec]
-        # Restore original DB
-        r select 9
-        set res
-    } {PONG}
+        test {WATCH is able to remember the DB a key belongs to} {
+            r select 5
+            r set x 30
+            r watch x
+            r select 1
+            r set x 10
+            r select 5
+            r multi
+            r ping
+            set res [r exec]
+            # Restore original DB
+            r select 9
+            set res
+        } {PONG}
+    }
 
     test {WATCH will consider touched keys target of EXPIRE} {
         r del x
